1.API文档解析服务平台/requirements.txt
Flask>=2.0.0
Flask-CORS>=3.0.0
requests>=2.25.0
PyYAML>=5.4.0
python-dotenv>=0.19.0